requirements.txt
ldap3
pyyaml
flask==1.0
pyOpenSSL==18.0.0
decorator==4.3.0
rfc5424-logging-handler
swagger-ui-bundle==0.0.2

# Use development version which works
#  with openapi v3
https://github.com/zalando/connexion/archive/dev-2.0.zip

